+++ src/Table/Actions/Presets/ArchiveRowAction.php
@@ -1,0 +1,45 @@
+<?php
+
+namespace Thinktomorrow\Chief\Table\Actions\Presets;
+
+use Thinktomorrow\Chief\ManagedModels\States\Actions\UpdateState;
+use Thinktomorrow\Chief\ManagedModels\States\State\GetPrimaryStateKeyOfModel;
+use Thinktomorrow\Chief\ManagedModels\States\State\StateMachine;
+use Thinktomorrow\Chief\Shared\ModelReferences\ModelReference;
+use Thinktomorrow\Chief\Table\Actions\Action;
+
+class ArchiveRowAction extends Action
+{
+    public static function makeDefault(string $resourceKey, ?string $stateKey = null, string $transitionKey = 'archive'): static
+    {
+        if (! $primaryStateKey = GetPrimaryStateKeyOfModel::get($resourceKey)) {
+            throw new \RuntimeException('ArchiveRowAction requires a primary state key to be defined on the model.');
+        }
+
+        $stateKey = $stateKey ?: $primaryStateKey;
+
+        return static::make('archive-state-row')
+            ->label('Archiveren')
+            ->variant('red')
+            ->prependIcon('<x-chief::icon.view />')
+            ->effect(function ($formData, $data) use ($resourceKey, $stateKey, $transitionKey) {
+
+                app(UpdateState::class)->handle(
+                    $resourceKey,
+                    ModelReference::fromString($data['item']),
+                    $stateKey,
+                    $transitionKey,
+                    []
+                );
+
+                return true;
+            })
+            ->notifyOnSuccess('Gearchiveerd')->notifyOnFailure('Er is iets misgegaan bij het archiveren van dit item.')
+            ->when(function ($component, $model) use ($stateKey, $transitionKey) {
+                $stateConfig = $model->getStateConfig($stateKey);
+                $stateMachine = StateMachine::fromConfig($model, $stateConfig);
+
+                return $stateMachine->can($transitionKey);
+            });
+    }
+}